Update project configuration and dependencies
The changes to the project's configuration include setting the module to "esnext" in tsconfig.json and modifying start commands for iOS and Android platforms in package.json. Several new dependencies were added, including axios, expo-file-system, expo-sqlite, and firebase. Furthermore, the color scheme in the Colors.ts file was updated and the version of some packages was bumped in package-lock.json.